+++ air/src/trace/mod.rs
@@ -42,7 +42,7 @@ pub const STACK_TRACE_RANGE: Range<usize> = range(STACK_TRACE_OFFSET, STACK_TRAC
 
 // Range check trace
 pub const RANGE_CHECK_TRACE_OFFSET: usize = STACK_TRACE_RANGE.end;
-pub const RANGE_CHECK_TRACE_WIDTH: usize = 3;
+pub const RANGE_CHECK_TRACE_WIDTH: usize = 2;
 pub const RANGE_CHECK_TRACE_RANGE: Range<usize> =
     range(RANGE_CHECK_TRACE_OFFSET, RANGE_CHECK_TRACE_WIDTH);
 
@@ -57,7 +57,7 @@ pub const TRACE_WIDTH: usize = CHIPLETS_OFFSET + CHIPLETS_WIDTH;
 // ------------------------------------------------------------------------------------------------
 
 //      decoder         stack       range checks      hasher         chiplets
-//    (3 columns)     (1 column)     (2 columns)    (1 column)      (1 column)
+//    (3 columns)     (1 column)     (1 column)    (1 column)      (1 column)
 // ├───────────────┴──────────────┴──────────────┴───────────────┴───────────────┤
 
 // Decoder auxiliary columns
@@ -74,7 +74,7 @@ pub const STACK_AUX_TRACE_RANGE: Range<usize> =
 
 // Range check auxiliary columns
 pub const RANGE_CHECK_AUX_TRACE_OFFSET: usize = STACK_AUX_TRACE_RANGE.end;
-pub const RANGE_CHECK_AUX_TRACE_WIDTH: usize = 2;
+pub const RANGE_CHECK_AUX_TRACE_WIDTH: usize = 1;
 pub const RANGE_CHECK_AUX_TRACE_RANGE: Range<usize> =
     range(RANGE_CHECK_AUX_TRACE_OFFSET, RANGE_CHECK_AUX_TRACE_WIDTH);